I18N: converting faq package to using the message catalog. Adding some danish translations. Adding popup javascript confirmation to deletion of FAQ:s and questions

+++ www/admin/index.adp
@@ -1,9 +1,9 @@
 <master>
 <property name="context">@context@</property>
-<property name="title">FAQ Admin</property>
+<property name="title">#faq.FAQ_Admin#</property>
 
 <if @faqs:rowcount@ eq 0>
- <i>There are no FAQs available.</i><p>
+ <i>#faq.lt_no_FAQs#</i><p>
 </if>
 
 <else>
@@ -11,12 +11,13 @@
   <multiple name=faqs>
    <li><a href="one-faq?faq_id=@faqs.faq_id@">@faqs.faq_name@</a>
 	( 
-	<a href="faq-edit?faq_id=@faqs.faq_id@">edit</a> |
-	<a href="faq-delete?faq_id=@faqs.faq_id@">delete</a> 
+	<a href="faq-edit?faq_id=@faqs.faq_id@">#faq.edit#</a> |
+	<a href="faq-delete?faq_id=@faqs.faq_id@" onclick="return confirm('#faq.lt_Are_you_sure_you_want#');">#faq.delete#</a> 
 	)
     </li>
   </multiple>
  </ul>
 </else>
 
-  <a href="faq-new">Create a new FAQ</a>
+  <a href="faq-new">#faq.Create_a_new_FAQ#</a>
+
